Revert "the merge between main and branch with db"
