--- app/requirements.txt
@@ -1,29 +1,0 @@
-# Core Flask and extensions
-Flask==2.3.3
-Flask-Login==0.6.3
-Flask-SQLAlchemy==3.1.1
-Flask-WTF==1.2.1
-Flask-Migrate==4.0.5
-
-# Database
-SQLAlchemy==2.0.23
-alembic==1.13.0  # Used by Flask-Migrate
-
-# Forms
-WTForms==3.1.1
-
-# Security
-Werkzeug==3.0.1
-
-# Web requirements
-Jinja2==3.1.3
-MarkupSafe==2.1.3
-itsdangerous==2.1.2  # Required for Flask sessions
-
-# Testing (recommended)
-pytest==7.4.3
-pytest-flask==1.3.0
-
-# Development tools (optional)
-flask-shell-ipython==0.5.1
-python-dotenv==1.0.0  # For environment variable management
