--- app/static/css/account.css
@@ -1,125 +1,0 @@
-.account-container {
-    max-width: 1000px;
-    margin: 0 auto;
-    padding: 20px;
-}
-
-.account-header {
-    margin-bottom: 30px;
-    border-bottom: 1px solid #eaeaea;
-    padding-bottom: 15px;
-}
-
-.account-email {
-    color: #666;
-    margin-top: -10px;
-}
-
-.account-actions {
-    margin-bottom: 25px;
-}
-
-.account-section {
-    margin-top: 40px;
-}
-
-.preset-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.preset-card {
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    padding: 15px;
-    background: #f9f9f9;
-    box-shadow: 0 2px 4px rgba(0,0,0,0.05);
-}
-
-.preset-card.shared {
-    background: #f0f7ff;
-    border-color: #cce0ff;
-}
-
-.preset-header {
-    margin-bottom: 15px;
-}
-
-.preset-header h3 {
-    margin: 0 0 5px 0;
-}
-
-.preset-date, .shared-by {
-    font-size: 0.8rem;
-    color: #777;
-    display: block;
-}
-
-.preset-actions {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 8px;
-}
-
-.btn {
-    padding: 8px 12px;
-    border-radius: 4px;
-    text-decoration: none;
-    border: none;
-    cursor: pointer;
-    font-size: 14px;
-    background: #eee;
-    color: #333;
-}
-
-.btn-sm {
-    padding: 5px 10px;
-    font-size: 12px;
-}
-
-.btn-primary {
-    background: #4CAF50;
-    color: white;
-}
-
-.btn-danger {
-    background: #f44336;
-    color: white;
-}
-
-.no-presets {
-    text-align: center;
-    padding: 40px;
-    background: #f9f9f9;
-    border-radius: 8px;
-    margin-top: 20px;
-}
-
-.modal {
-    display: none;
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: rgba(0,0,0,0.5);
-    align-items: center;
-    justify-content: center;
-    z-index: 1000;
-}
-
-.modal-content {
-    background: white;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 2px 10px rgba(0,0,0,0.1);
-}
-
-.modal-actions {
-    margin-top: 20px;
-    display: flex;
-    justify-content: flex-end;
-    gap: 10px;
-}
