+++ app/static/css/presets.css
@@ -1,3 +1,4 @@
+
 /* Container */
 .presets-container {
     max-width: 1200px;
@@ -163,22 +164,17 @@
 
 /* ======== EMPTY STATES ======== */
 .empty-state {
-    background-color: #f9f9f9;
+    background-color: #222;
     border-radius: 8px;
-    padding: 30px;
+    padding: 3rem 2rem;
     text-align: center;
     max-width: 600px;
-    margin: 20px 0;
+    margin: 0 auto;
 }
 
 .empty-state h3 {
     margin-bottom: 1rem;
     color: #fff;
-}
-
-.empty-state h4 {
-    margin-bottom: 10px;
-    color: #555;
 }
 
 .empty-state p {
@@ -351,42 +347,4 @@
         width: 100%;
         text-align: center;
     }
-}
-
-body {
-    background-color: #121212; /* Dark background for the entire page */
-    font-family: Arial, sans-serif;
-}
-
-/* Community preset styles */
-.section-header {
-    margin-top: 30px;
-    margin-bottom: 15px;
-    padding-bottom: 10px;
-    border-bottom: 1px solid #eaeaea;
-}
-
-.section-header h3 {
-    font-size: 1.2rem;
-    margin-bottom: 5px;
-}
-
-.section-header p {
-    color: #666;
-    margin: 0;
-}
-
-.shared-section {
-    margin-top: 40px;
-}
-
-.preset-creator {
-    font-size: 0.85rem;
-    color: #666;
-    margin-top: 4px;
-}
-
-/* Style differences for shared presets */
-.shared-preset {
-    border-left: 3px solid #4a90e2; /* Blue indicator for shared presets */
 }
