+++ app/templates/account.html
@@ -1,81 +1,10 @@
+<!DOCTYPE html>
 {% extends "body.html" %}
-
 {% block content %}
-<head>
-    <link rel="stylesheet" href="/static/css/account.css">
-    <script src = "/static/js/account.js"></script>
-</head>
-<div class="account-container">
-    <div class="account-header">
-        <h1>Welcome, {{ current_user.fname }} {{ current_user.lname }}</h1>
-        <p class="account-email">{{ current_user.email }}</p>
-    </div>
-    
-    <div class="account-actions">
-        <a href="{{ url_for('rgb') }}" class="btn btn-primary">Create New Preset</a>
-    </div>
-    
-    <div class="account-content">
-        <h2>Your Saved Presets</h2>
-        
-        {% if uploads %}
-            <div class="preset-grid">
-                {% for preset in uploads %}
-                <div class="preset-card">
-                    <div class="preset-header">
-                        <h3>{{ preset.preset_name }}</h3>
-                    </div>
-                    
-                    <div class="preset-actions">
-                        <a href="{{ url_for('view_preset', preset_id=preset.id) }}" class="btn btn-sm">view</a>
-                        <a href="{{ url_for('edit_preset', preset_id=preset.id) }}" class="btn btn-sm">edit</a>
-                        <a href="{{ url_for('share_preset', preset_id=preset.id) }}" class="btn btn-sm btn-share">share</a>
-                        <button onclick="confirmDelete({{ preset.id }})" class="btn btn-sm btn-danger" id = 'confirmDeleteBtn'>delete</button>
-                    </div>
-                </div>
-                {% endfor %}
-            </div>
-        {% else %}
-            <div class="no-presets">
-                <p>You haven't saved any presets yet.</p>
-                <p>Go to the <a href="{{ url_for('rgb') }}">RGB editor</a> to create your first preset!</p>
-            </div>
-        {% endif %}
-    </div>
-    
-    <div class="account-section">
-        <h2>Shared With You</h2>
-        {% if shared_with_me %}
-            <div class="preset-grid">
-                {% for shared in shared_with_me %}
-                <div class="preset-card shared">
-                    <div class="preset-header">
-                        <h3>{{ shared.preset.preset_name }}</h3>
-                        <span class="shared-by">Shared by: {{ shared.shared_by.fname }} {{ shared.shared_by.lname }}</span>
-                    </div>
-                    
-                    <div class="preset-actions">
-                        <a href="{{ url_for('view_preset', preset_id=shared.preset_id) }}" class="btn btn-sm">View</a>
-                        <a href="{{ url_for('load_preset', preset_id=shared.preset_id) }}" class="btn btn-sm btn-primary">Load in Editor</a>
-                    </div>
-                </div>
-                {% endfor %}
-            </div>
-        {% else %}
-            <p>No presets have been shared with you yet.</p>
-        {% endif %}
-    </div>
-</div>
-
-<!-- Delete Confirmation Modal -->
-<div id="deleteModal" class="modal">
-    <div class="modal-content">
-        <h2>Confirm Deletion</h2>
-        <p>Are you sure you want to delete this preset? This action cannot be undone.</p>
-        <div class="modal-actions">
-            <button id="confirmDeleteBtn" class="btn btn-danger">Delete</button>
-            <button onclick="closeModal()" class="btn">Cancel</button>
-        </div>
-    </div>
+<div class="container">
+    <h1>Welcome, </h1>
+    <p>Your activity history:</p>
+    <ul>
+    </ul>
 </div>
 {% endblock %}
