+++ app/templates/body.html
@@ -18,25 +18,12 @@
             <a href="{{ url_for('home') }}">Scyra</a>
         </h1>
         <nav class="navbar">
-            <a href="{{ url_for('home') }}">Home page</a>
+            <a href="{{ url_for('home') }}">Home</a>
             <a href="{{ url_for('rgb') }}">RGB</a>
             <a href="{{ url_for('presets') }}">Presets</a>
             <a href="{{ url_for('about') }}">About</a>
-            <!-- Conditional navigation links -->
-            {% if current_user.is_authenticated %}
-            <!-- Show when logged in -->
-            <div class="user-dropdown">
-                <span class="user-greeting">Hello, {{ current_user.fname }}</span>
-                <div class="dropdown-content">
-                    <a href="{{ url_for('account') }}">Account</a>
-                    <a href="{{ url_for('logout') }}">Logout</a>
-                </div>
-            </div>
-            {% else %}
-                <!-- Show when logged out -->
-                <a href="{{ url_for('login') }}">Login</a>
-                <a href="{{ url_for('register') }}">Register</a>
-            {% endif %}
+            <a href="{{ url_for('login') }}">Login</a>
+            <a href="{{ url_for('account') }}">Account</a>
         </nav>
     </header>
