+++ app/templates/register.html
@@ -5,21 +5,12 @@
 {% block content %}
 <h2>Register</h2>
 <form id="register-form" action="{{ url_for('register') }}" method="POST">
-    <label for="email">Email:</label>
-    <input type="email" id="email" name="email" required>
-    
-    <label for="fname">First Name:</label>
-    <input type="text" id="fname" name="fname" required>
-    
-    <label for="lname">Last Name:</label>
-    <input type="text" id="lname" name="lname" required>
-    
+    <label for="new_username">Username:</label>
+    <input type="text" id="new_username" name="new_username" required>
     <label for="new_password">Password:</label>
     <input type="password" id="new_password" name="new_password" required>
-    
     <label for="confirm_password">Confirm Password:</label>
     <input type="password" id="confirm_password" name="confirm_password" required>
-    
     <button type="submit">Register</button>
 </form>
 <p id="error-message" style="color: red; display: none;">Passwords do not match!</p>
